Improve tag name determination
Tries to ensure that there’s only ever one h1 tag on a page
