+++ assets/src/helpers.js
@@ -6,6 +6,12 @@
 import uuid from 'uuid/v4';
 
 /**
+ * WordPress dependencies
+ */
+import { count } from '@wordpress/wordcount';
+import { _x } from '@wordpress/i18n';
+
+/**
  * Internal dependencies
  */
 import { ALLOWED_CHILD_BLOCKS, ALLOWED_TOP_LEVEL_BLOCKS, BLOCK_TAG_MAPPING } from './constants';
@@ -216,4 +222,56 @@ export const addAMPExtraProps = ( props, blockType, attributes ) => {
 		...props,
 		...ampAttributes,
 	};
+};
+
+// Todo: Make these customizable?
+const H1_FONT_SIZE = 40;
+const H2_FONT_SIZE = 24;
+const H1_TEXT_LENGTH = 4;
+const H2_TEXT_LENGTH = 10;
+
+/*
+ * translators: If your word count is based on single characters (e.g. East Asian characters),
+ * enter 'characters_excluding_spaces' or 'characters_including_spaces'. Otherwise, enter 'words'.
+ * Do not translate into your own language.
+ */
+const wordCountType = _x( 'words', 'Word count type. Do not translate!', 'amp' );
+
+/**
+ * Determines the HTML tag name that should be used given on the block's attributes.
+ *
+ * Font size takes precedence over text length as it's a stronger signal for semantic meaning.
+ *
+ * @param {Object}  attributes Block attributes.
+ * @param {boolean} canUseH1   Whether an H1 tag is allowed.
+ *
+ * @return {string} HTML tag name. Either p, h1, or h2.
+ */
+export const getTagName = ( attributes, canUseH1 ) => {
+	const { fontSize, customFontSize, positionTop } = attributes;
+
+	// Elements positioned that low on a page are unlikely to be headings.
+	if ( positionTop > 80 ) {
+		return 'p';
+	}
+
+	if ( 'huge' === fontSize || ( customFontSize && customFontSize > H1_FONT_SIZE ) ) {
+		return canUseH1 ? 'h1' : 'h2';
+	}
+
+	if ( 'large' === fontSize || ( customFontSize && customFontSize > H2_FONT_SIZE ) ) {
+		return 'h2';
+	}
+
+	const textLength = count( attributes.content, wordCountType, {} );
+
+	if ( H1_TEXT_LENGTH >= textLength ) {
+		return canUseH1 ? 'h1' : 'h2';
+	}
+
+	if ( H2_TEXT_LENGTH >= textLength ) {
+		return 'h2';
+	}
+
+	return 'p';
 };
